Moved address, memory and pose services to the services directory
