+++ Anamnesis/Actor/Posing/SkeletonEntity.cs
@@ -3,7 +3,6 @@
 
 namespace Anamnesis.Actor.Posing;
 
-using Anamnesis.Actor;
 using Anamnesis.Core;
 using Anamnesis.Memory;
 using Anamnesis.Services;
