+++ Anamnesis/Files/PoseFile.cs
@@ -3,11 +3,11 @@
 
 namespace Anamnesis.Files;
 
-using Anamnesis.Actor;
 using Anamnesis.Core;
 using Anamnesis.Core.Extensions;
 using Anamnesis.Memory;
 using Anamnesis.Posing;
+using Anamnesis.Services;
 using Serilog;
 using System;
 using System.Collections.Generic;
