+++ Anamnesis/Serialization/SerializerService.cs
@@ -10,7 +10,6 @@ using System.IO;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 
-// TODO: Move this file to Services folder.
 public class SerializerService : ServiceBase<SerializerService>
 {
 	public static JsonSerializerOptions Options = new JsonSerializerOptions();
