+++ Anamnesis/Services/AddressService.cs
@@ -1,7 +1,7 @@
 ﻿// © Anamnesis.
 // Licensed under the MIT license.
 
-namespace Anamnesis.Core.Memory;
+namespace Anamnesis.Services;
 
 using Anamnesis.Core;
 using Anamnesis.Memory;
@@ -10,8 +10,9 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using XivToolsWpf;
 
-// TODO: Move this to the services directory.
-
+/// <summary>
+/// A service that handles the scanning and resolution of memory addresses from the game process.
+/// </summary>
 #pragma warning disable SA1027, SA1025
 public class AddressService : ServiceBase<AddressService>
 {
@@ -50,6 +51,9 @@ public class AddressService : ServiceBase<AddressService>
 	public static IntPtr KineDriverRotation { get; private set; }
 	public static IntPtr KineDriverScale { get; private set; }
 
+	/// <summary>
+	/// Gets the camera address from the game process.
+	/// </summary>
 	public static IntPtr Camera
 	{
 		get
@@ -63,6 +67,9 @@ public class AddressService : ServiceBase<AddressService>
 		}
 	}
 
+	/// <summary>
+	/// Gets the weather address from the game process.
+	/// </summary>
 	public static IntPtr Weather
 	{
 		get
@@ -82,6 +89,9 @@ public class AddressService : ServiceBase<AddressService>
 		}
 	}
 
+	/// <summary>
+	/// Gets the GPose weather address from the game process.
+	/// </summary>
 	public static IntPtr GPoseWeather
 	{
 		get
@@ -97,6 +107,9 @@ public class AddressService : ServiceBase<AddressService>
 		}
 	}
 
+	/// <summary>
+	/// Gets the GPose camera address from the game process.
+	/// </summary>
 	public static IntPtr GPoseCamera
 	{
 		get
